Build objects for the protocol library in the same subdirs as their sources
- Build static libraries in the classes and support subdirs, and then
merge them into a top-level static library in the protocol directory.
Switch to using libtool libraries, since Automake requires us to use libtool
for this.

- Remove AC_DISABLE_STATIC from configure.ac, since libtool can't make
static libraries if this directive is there.  This means that we'll always
build both static and shared versions of libvdisk, and then delete the
static version during the install process, since libtool doesn't provide
a per-object option to disable static libraries.

- Add .gitignore files for the protocol library