Have genflow unconditionally create output files
- Have genflow always create its output files, even if the existing files
are identical to the new ones.  Since Makefile.am now runs genflow whenever
the script itself is newer than its output, the targets made by the script
must always be updated by it or else the script will be run on every
invocation of Make if its output for only one file has been modified.  We
could work around this with a stamp file (the way compile.pl does) but it
doesn't seem worth it for only two output files.