libprotocol: Add -fcompound-names to asn1c options
By default, asn1c creates substructures within a top-level type as
follows:

struct class {
	...
} class;

The structure type doesn't even need to be named in this case, but
asn1c does it anyway.  So, if there are different top-level types, both
of which have e.g. a "class" field which is a SET OF things, this will
generate compile errors.  So we add -fcompound-names to mangle the struct
names as follows:

struct TopLevelType__class {
	...
} class;

This should be harmless since we don't care about the structure type name
anyway.