minirpc: Only export functions tagged "exported" (with newer gcc)
The minirpc library has a number of functions which are called across file
boundaries (so that they can't be static) but which are not intended to
be called (or overridden with LD_PRELOAD) from outside the library.  Newer
versions of gcc support setting symbol visibility outside the linked
library.

- Add configure test for -fvisibility=hidden compiler support

- If the compiler supports visibility, compile minirpc with no symbols
exported by default

- Define "exported" symbol attribute which exports the symbol