minirpc: Force in-order processing of events from a connection
- Maintain a two-stage hierarchy of pending events: the conn set tracks
connections with pending events, and each connection tracks pending
messages.

- Prevent further messages from a connection from being processed while the
connection's head-of-queue is processing (i.e., plug the connection queue).
This allows the application to prevent races when a connection makes a state
transition (e.g. unauthenticated -> authenticated) and multiple worker
threads are processing messages from the same conn set.  When the first
event is done processing, the connection is requeued at the end of the
conn-set queue (round-robin connection scheduling).  If the application
request function returns MINIRPC_DEFER, automatically unplug the connection
(since otherwise the plugging would make server async into per-connection
synchronous).

- Allow application request handlers to unplug the connection (e.g. if
the handler is long-running and it's safe to continue processing messages)
by calling mrpc_unplug_event(), which returns an error if the provided
message is not currently plugging the connection