Port mrpc_listen() to APR socket functions
- Port mrpc_listen() to APR socket functions.  Change the prototype to
return an error code and to pass the number of bound sockets as an output
parameter, rather than the other way around.  Change the error code from a
string to an apr_status_t, since that's essentially equivalent to an OS
error code but more expressive.  (We originally returned a string because
getaddrinfo error codes collided with errno-style codes, but this approach
was obviously suboptimal. It seems better to return an apr_status_t directly
than to destructively map it to an OS error.)

- Remove setsockoptval() and set_nonblock(), since their last users are gone