Distinguish mrpc_conn_close() and conn_kill() shutdown paths
- Add shutdown flag field to mrpc_connection struct, with its own lock.
Use it to track the various things that need to happen during shutdown.

- Provide two shutdown paths: when the application requests that a conn be
closed, drop most incoming messages (but don't close the fd until outgoing
messages are sent), and when the system decides to shut down a conn, drop
all outgoing messages and close the fd.  If both minirpc and the app have
asked to close the conn, we drop both incoming and outgoing messages.
When we drop incoming messages, we discard requests and IO errors and
return MINIRPC_NETWORK_FAILURE for replies.

- When freeing the pending hash, just release its members without waking its
callbacks.  Provide a separate function for waking pending callbacks. This
is for safety, to make sure that no callbacks are accidentally invoked
during the mrpc_conn_free() after EVENT_DISCONNECT (when the application is
no longer equipeed to deal with them).  Wake pending callbacks after the
send and event queues have been flushed, but before EVENT_DISCONNECT.

- In the send_message() error path, free the mrpc_message outside the locks

- Move forcible event destruction into a helper function