Have mrpc_conn_close() return -EALREADY after first call
- If mrpc_conn_close() is called on a conn after a conn_kill(), be sure
to squash the event queue (since that's what the application expects).
If the fd is already closed, then assume that EVENT_Q_SHUTDOWN has already
been queued.

- In order to prevent deadlocks between two event threads each calling
mrpc_conn_close() on the same connection, have mrpc_conn_close() decline to
provide the no-events-running guarantee if it is called a second time on a
particular conn.  Return -EALREADY in this case so that the application can
distinguish it.

- Move SHUT_SQUASH_EVENTS handling from conn_start_shutdown() to
mrpc_conn_close()

- Have conn_start_shutdown() return void