Fix mrpc_conn_set_destroy() race with mrpc_dispatch_{one,all}()
There is a race involving mrpc_conn_set_destroy() if the application is
using mrpc_dispatch_one()/mrpc_dispatch_all() to dispatch its events: the
conn set cannot be destroyed until all pending events are handled, but it's
not safe to call the dispatch functions as the conn set is being destroyed.
As a result, the app can't know exactly when to stop calling the dispatch
functions.  Therefore:

- Change mrpc_dispatch_one()/mrpc_dispatch_all() to return an error code
rather than whether an event was handled (for the former) or the number
of events that were handled (for the latter).  Return -EAGAIN from the
former if no event was handled.

- Add exported mrpc_dispatcher_add()/mrpc_dispatcher_remove() functions,
which get an events_threads reference the same way mrpc_dispatch_loop()
does.  The application is expected to call these from any thread which is
going to be dispatching events.  This includes dispatching via
mrpc_dispatch_loop(), which no longer gets an events_threads reference
directly.  Add mrpc_dispatcher_add()/mrpc_dispatcher_remove() calls to the
start routine used by mrpc_start_dispatch_thread().

- Return -ESHUTDOWN from mrpc_dispatch_*() when the application should no
longer call the dispatch functions for this set.  The app is expected to
call mrpc_dispatcher_remove() at this point to drop the set ref.

- Fix pollset_free(NULL) in mrpc_dispatch_loop() error path