Do not duplicate the event fd in mrpc_get_event_fd()
- mrpc_get_event_fd(): Return the fd obtained from selfpipe_fd(), rather
than a duplicate of it.  We can do this because the mrpc_conn_set_destroy()
semantics make it reasonable to close the fd out from under the application,
since the app knows exactly when to stop polling on it.  This change also
reduces the number of fds we use for pipes.

- Have mrpc_get_event_fd() return the fd directly, rather than through an
out argument