Make the event fd readable when the conn set is being shut down
- Make the event fd readable when the conn set is being shut down.  We need
this because applications using mrpc_get_event_fd() don't have access to the
shutdown pipe and need to be woken when a shutdown is in progress.

- Move the events_notify_pipe logic to a central helper function

- Don't poll on shutdown_pipe in mrpc_dispatch_loop(), since it's no longer
necessary

- Improve performance of selfpipe code at the expense of safeguards.  Now we
only write to the pipe if it isn't already set, we only read from the pipe
if it is set, and we require that only one byte be waiting in the pipe when
emptying it.  Applications which try to read the event pipe are already
going to break things, and it's documented that they shouldn't do that, so
we shouldn't need to work around them.