Fix mrpc_unplug_event()/EVENT_DISCONNECT race
- The connection shutdown sequencing did not take into account the fact
that an mrpc_unplug_event() call from a running event could cause
EVENT_DISCONNECT to complete while the first event still needed the
connection handle.  Because events in flight are not reflected in the
pending-event queue, convert the shutdown sequencer from using queue
operations (EVENT_Q_SHUTDOWN) to reference counting.  Take a conn
reference as soon as an event is queued for the conn, and release it
when the event has been completely handled.  We now don't even queue
the disconnect event until all other event handlers have finished.