Properly handle messages longer than msg_max_buf_len
- Add recv_error member to struct mrpc_message.  If nonzero, this error code
is returned by unformat_{request,reply}() in preference to the error code
embedded in the message header.  Set recv_error in process_incoming_header()
if unserializing the miniRPC header fails, or if the data length specified
in the header exceeds msg_max_buf_len.

- Add STATE_INVALID to try_read_conn() state machine, and enter it when
process_incoming_header() produces a nonzero recv_error.  When in INVALID
state, read the miniRPC payload out of the connection and drop it on the
floor, so that the connection stays in sync for subsequent messages.  Add
per-conn-set trash buffer for this purpose.  Convert the receive counter
into a down-counter so that we're not dependent on buffer offsets when using
the trash buffer.

- Make process_incoming_header() return void, since the error code can be
read out of conn->recv_msg->recv_error

- Remove unused recv_length member from struct mrpc_connection

- Add min() and max() macros to internal.h