Various cleanups
- Remove unnecessary test for NULL request function in dispatch_request(),
since the receiver_request_info function should have already handled the
case in which there are no receiver operations for this role.  Add an
assert() to replace it.

- When squashing events in dispatch_event(), remember to unplug events
before destroying them

- When handling a reply with a command mismatch or with both an error code
and a payload, treat the problem as a receive error and wake the waiting
sender, rather than treating it as an unmatched reply