Properly handle accept() failures in try_accept()
- Add simple pollset mechanism which allows one-shot timers to be set
against individual fds.  When the timer expires, the specified method
pointer is called.  The mechanism non-racily cleans up pending timers in the
pollset_del() path, but does not otherwise provide a way to modify or cancel
pending timers, since that functionality is not needed at the moment.

- If try_accept() encounters an error in accept(), disable POLLSET_READABLE
on the listening fd and register a timer to re-enable it later.  Add
accept_backoff mrpc_config parameter which specifies the amount of time
to wait before re-enabling accept().  This fixes busy-loops occurring when
the process runs out of file descriptors.