Garbage-collect unreplied message handles in mrpc_conn_free()
- When the disconnect event arrives, the application may hold mrpc_message
handles to requests for which it returned MINIRPC_PENDING.  Even if it tries
to call send_async_reply_error on those handles, the call won't succeed
(because the connection is dead) and so the handles won't be cleaned up.
So, we need to clean up the remaining messages ourselves.  Add a list of
messages allocated against a connection, and make sure to free them all from
mrpc_conn_free().