Add mrpc_conn_ref()/mrpc_conn_unref()
- Add mrpc_conn_ref()/mrpc_conn_unref() API analogous to conn set API.
mrpc_conn_close() no longer implies freeing the conn handle; its refcount
must drop to zero first.  This allows the application to safely make a
series of synchronous calls without fear of a bad pointer dereference if
the other end closes the connection.  A connected handle holds an additional
reference itself, so an application can drop the user-visible refcount to
zero and the connection handle will be freed as soon as the disconnect
method returns.

- Have mrpc_conn_close() return ENOTCONN on unconnected handle, and have
no other effect

- mrpc_conn_close(): Move sequence_lock unlock to common return path

- Update documentation and test suite

- test_shutdown: Remove can_free semaphore, since we no longer need to
block the disconnect function until we're done making calls on the conn