Fix test_errors segfault on 64-bit systems
- On 64-bit systems, xdr_int() tries to dereference its argument directly,
rather than just passing it to the proper XDR stream.  This means that it
isn't safe for format_message() to assume that serialize_common() will catch
NULL data arguments, since we pass the argument through the xdrlen stream
before calling serialize_common().  So, add a check to format_message()
which errors out on NULL data pointers if the xdr_proc is not xdr_void.

- Remove NULL data pointer test from serialize_common(), since it's now
redundant