API CHANGE: Change prototypes for mrpc_connect() and mrpc_listen()
Change mrpc_connect() and mrpc_listen() prototypes to allow greater
generality w.r.t. the underlying transport:

- Rename "port" parameter to "service", and make it a string rather than an
unsigned int.  This matches the getaddrinfo() interface, removes the
assumption that port numbers are equal across protocols, allows service
names to be used as well as numbers, and removes the need to hardcode
sockaddr-to-port-number translations for each protocol.

- Add address-family parameter.  If AF_UNSPEC, the old behavior will be
used; otherwise, sockets will be restricted to the specified address
family.  This gives the application more control over the connection types
in use.  To eliminate ambiguity, require the family to be specified
when calling mrpc_listen() with *service == NULL.

- Move getaddrinfo error code translator to its own helper function

- Update documentation, tests, and examples