minirpcgen: Put output files in current directory if -o not specified
- If the output basepath was left implicit, output files would be produced
in the directory of the input .mx file, rather than in the current
directory.  This made it impossible to leave off the -o option when using
Automake, since it would break VPATH builds.  Change this behavior so that
output files are always created in the current directory unless requested
otherwise with an explicit -o option.